# By asiainfo (2) and 王坜 (1)
# Via asiainfo (2) and others
* 'appStore' of github.com:nicolastinkl/AI2020OS:
  fix share
  OK commit with tinkl
  update

+++ Veris/Business/CustomerCenter/View/BuyerViews/AIServiceCoverage.m
@@ -103,7 +103,7 @@
         AIServiceLabel *label = [[AIServiceLabel alloc] initWithFrame:frame title:title type:AIServiceLabelTypeSelection isSelected:model.isSelected];
         label.delegate = self;
         label.backgroundColor = [self makeColorsAtIndex:i];
-        label.selectionImageView.image = [UIImage imageNamed:[NSString stringWithFormat:@"Coverage%ld", i]];
+        label.selectionImageView.image = [UIImage imageNamed:[NSString stringWithFormat:@"Coverage%ld", (long)i]];
 
         [self addSubview:label];
